move serveHTTP method from separate file
